Moved copy() to new memory sub-module
