+++ heat/core/tests/test_operations.py
@@ -376,19 +376,6 @@ class TestOperations(unittest.TestCase):
         with self.assertRaises(ValueError):
             ht.argmin(data, axis=-4)
 
-    def test_copy(self):
-        tensor = ht.ones(5)
-        copied = tensor.copy()
-
-        # test identity inequality and value equality
-        self.assertIsNot(tensor, copied)
-        self.assertIsNot(tensor._Tensor__array, copied._Tensor__array)
-        self.assertTrue((tensor == copied)._Tensor__array.all())
-
-        # test exceptions
-        with self.assertRaises(TypeError):
-            ht.copy('hello world')
-
     def test_transpose(self):
         # vector transpose, not distributed
         vector = ht.arange(10)
